Set Up Login with test and userExtractor for certain routes middlewares.

+++ controllers/blogs.js
@@ -1,5 +1,10 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const { SECRET } = require('../utils/config')
+const jwt = require('jsonwebtoken')
+const middlewares = require('../utils/middlewares')
+
 /* Working with promises
 blogsRouter.get('/', (request, response) => {
   Blog
@@ -9,9 +14,13 @@ blogsRouter.get('/', (request, response) => {
     })
 })
 */
+
 // Working with async/await
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('author', { username: 1, name: 1, id: 1 })
+
   response.json(blogs)
 })
 
@@ -36,21 +45,70 @@ blogsRouter.post('/', (request, response) => {
 })
 */
 // Working with async/await
-blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
+blogsRouter.post('/', middlewares.userExtractor, async (request, response) => {
+  const blog = new Blog(request.body)  
+/*
+  const users = await User.find({});
+  console.log('Available users for blog authorship:', users);
+  const user = users[0];
+ */
+  //const token = request.token
+
+  //if (!token) {
+  //  return response.status(401).json({ error: 'token missing' })
+  //}
+  
+  //const decodedToken = jwt.verify(token, SECRET)
+  //if (!token || !decodedToken.id) {
+  //  return response.status(401).json({ error: 'token missing or invalid' })
+  //}
+
+  //const user = await User.findById(decodedToken.id)
+  const user = request.user
+
+  if (!user) {
+    return response.status(401).json({ error: 'user missing or invalid'})
+  }
+
+  blog.author = user;
+
   // Validate the blog before saving
   try {
     await blog.validate()
   } catch (error) {
     return response.status(400).json({ error: error.message })
   }
-
+  //console.log('new blog to insert (after validation):', blog);
   const savedBlog = await blog.save()
 
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middlewares.userExtractor, async (request, response) => {
+  //const token = request.token
+  
+  //if (!token) {
+  //  return response.status(401).json({ error: 'token missing' })
+  //}
+
+  //const decodedToken = jwt.verify(token, SECRET)
+
+  //if (!token || !decodedToken.id) {
+  //    return response.status(401).json({ error: 'token missing or invalid' })
+  //}
+  const user = request.user
+  //console.log('user request', user)
+  const db_blog = await Blog.findById(request.params.id)
+  
+  //console.log('db_blog to delete', db_blog)
+  if (!db_blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+  //console.log('author', db_blog.author, 'id', db_blog.author.toString())
+  if ( db_blog.author.toString() !== user ) {
+    return response.status(403).json({ error: 'only the author can delete this blog' })
+  }
+
   try {
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
